+++ src/main/java/com/reefbot/dto/slot/SlotSpinResponse.java
@@ -34,8 +34,8 @@ public record SlotSpinResponse(
         // VIP
         String         vipTier,
         long           vipLifetimeWager,
-        int            vipPeriodNetLoss,
-        int            vipEstimatedCashback,
+        long           vipPeriodNetLoss,
+        long           vipEstimatedCashback,
         Long           vipNextTierThreshold,
         String         vipNextCashbackDate
 ) {
@@ -45,7 +45,7 @@ public record SlotSpinResponse(
         return new SlotSpinResponse(
                 code, null, null, 0, 0, 0,
                 false, false, 0, 1, List.of(), 0, 0,
-                null, 0L, 0, 0, null, null);
+                null, 0L, 0L, 0L, null, null);
     }
 
     public static SlotSpinResponse ok(
@@ -62,8 +62,8 @@ public record SlotSpinResponse(
     ) {
         var tier     = player.getVipTier();
         var next     = tier.next();
-        int loss     = Math.max(0, player.getVipPeriodNetLoss());
-        int cashback = (int)(loss * tier.getCashbackRate());
+        long loss     = Math.max(0L, player.getVipPeriodNetLoss());
+        long cashback = (long)(loss * tier.getCashbackRate());
 
         var ss = player.getSlotState();
         return new SlotSpinResponse(
